feat: Add protobuf support for WebSocket messages and refactor WebSocket handling
- Updated `pyproject.toml` to include `grpcio-tools` and `protobuf` dependencies.
- Added generated protobuf files for WebSocket messages in `stackchan_server/generated_protobuf`.
- Implemented `protobuf_ws.py` for encoding and parsing WebSocket messages.
- Refactored `speak.py` to utilize new protobuf message encoding for audio WAV messages.
- Updated `ws_proxy.py` to handle protobuf messages and removed legacy struct-based message handling.
- Enhanced error handling for invalid protobuf messages in WebSocket communication.

+++ firmware/include/listening.hpp
@@ -53,7 +53,7 @@ private:
   size_t ring_read_ = 0;
   size_t ring_available_ = 0;
 
-  uint16_t seq_counter_ = 0;
+  uint32_t seq_counter_ = 0;
   bool streaming_ = false;
   bool events_registered_ = false;
